snmp: fix GC timer in snmp_note_store
in deactivate_timer/1 the new state of the timer is not set correctly,
instead the timeout value set to 'false', resulting in a "crash loop"
when the snmp note store process tries to restart the timer process.

the incorrect activation state also leads to an ETS memory leak:
an expired note will not get deleted except when
snmp_note_store:get_note/2 is called with its key - the GC mechanism
does not work because a deactivated timer will never be activated again.